api update

+++ testing/test_full.py
@@ -141,6 +141,12 @@ def test_constructTrainingCSV():
                 i += 1
 
 
+def sanitize_outputDir():
+    if os.path.isdir(outputDir):
+        shutil.rmtree(outputDir)  # overwrite previous results
+    Path(outputDir).mkdir(parents=True, exist_ok=True)
+
+
 def test_train_segmentation_rad_2d(device):
     print("Starting 2D Rad segmentation tests")
     # read and parse csv
@@ -161,8 +167,7 @@ def test_train_segmentation_rad_2d(device):
         parameters["model"]["architecture"] = model
         parameters["nested_training"]["testing"] = -5
         parameters["nested_training"]["validation"] = -5
-        shutil.rmtree(outputDir)  # overwrite previous results
-        Path(outputDir).mkdir(parents=True, exist_ok=True)
+        sanitize_outputDir()
         TrainingManager(
             dataframe=training_data,
             outputDir=outputDir,
@@ -192,8 +197,7 @@ def test_train_segmentation_sdnet_rad_2d(device):
     parameters["model"]["num_channels"] = 1
     parameters["model"]["architecture"] = "sdnet"
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
-    shutil.rmtree(outputDir)  # overwrite previous results
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
@@ -228,8 +232,7 @@ def test_train_segmentation_rad_3d(device):
         parameters["model"]["architecture"] = model
         parameters["nested_training"]["testing"] = -5
         parameters["nested_training"]["validation"] = -5
-        shutil.rmtree(outputDir)  # overwrite previous results
-        Path(outputDir).mkdir(parents=True, exist_ok=True)
+        sanitize_outputDir()
         TrainingManager(
             dataframe=training_data,
             outputDir=outputDir,
@@ -263,8 +266,7 @@ def test_train_regression_rad_2d(device):
         parameters["model"]["architecture"] = model
         parameters["nested_training"]["testing"] = -5
         parameters["nested_training"]["validation"] = -5
-        shutil.rmtree(outputDir)  # overwrite previous results
-        Path(outputDir).mkdir(parents=True, exist_ok=True)
+        sanitize_outputDir()
         TrainingManager(
             dataframe=training_data,
             outputDir=outputDir,
@@ -294,8 +296,7 @@ def test_train_brainage_rad_2d(device):
     parameters["scaling_factor"] = 1
     parameters["model"]["architecture"] = "brain_age"
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
-    shutil.rmtree(outputDir)  # overwrite previous results
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
@@ -327,8 +328,7 @@ def test_train_regression_rad_3d(device):
         parameters["model"]["architecture"] = model
         parameters["nested_training"]["testing"] = -5
         parameters["nested_training"]["validation"] = -5
-        shutil.rmtree(outputDir)  # overwrite previous results
-        Path(outputDir).mkdir(parents=True, exist_ok=True)
+        sanitize_outputDir()
         TrainingManager(
             dataframe=training_data,
             outputDir=outputDir,
@@ -361,8 +361,7 @@ def test_train_classification_rad_2d(device):
         parameters["model"]["architecture"] = model
         parameters["nested_training"]["testing"] = -5
         parameters["nested_training"]["validation"] = -5
-        shutil.rmtree(outputDir)  # overwrite previous results
-        Path(outputDir).mkdir(parents=True, exist_ok=True)
+        sanitize_outputDir()
         TrainingManager(
             dataframe=training_data,
             outputDir=outputDir,
@@ -394,8 +393,7 @@ def test_train_classification_rad_3d(device):
         parameters["model"]["architecture"] = model
         parameters["nested_training"]["testing"] = -5
         parameters["nested_training"]["validation"] = -5
-        shutil.rmtree(outputDir)  # overwrite previous results
-        Path(outputDir).mkdir(parents=True, exist_ok=True)
+        sanitize_outputDir()
         TrainingManager(
             dataframe=training_data,
             outputDir=outputDir,
@@ -425,7 +423,7 @@ def test_inference_classification_rad_3d(device):
     # loop through selected models and train for single epoch
     model = all_models_regression[0]
     parameters["model"]["architecture"] = model
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
@@ -489,7 +487,7 @@ def test_inference_classification_with_logits_single_fold_rad_3d(device):
     # loop through selected models and train for single epoch
     model = all_models_regression[0]
     parameters["model"]["architecture"] = model
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
@@ -530,7 +528,7 @@ def test_inference_classification_with_logits_multiple_folds_rad_3d(device):
     # loop through selected models and train for single epoch
     model = all_models_regression[0]
     parameters["model"]["architecture"] = model
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
@@ -584,7 +582,6 @@ def test_scheduler_classification_rad_2d(device):
             reset=True,
         )
 
-    shutil.rmtree(outputDir)
     print("passed")
 
 
@@ -622,7 +619,6 @@ def test_optimizer_classification_rad_2d(device):
             reset=True,
         )
 
-    shutil.rmtree(outputDir)
     print("passed")
 
 
@@ -647,8 +643,7 @@ def test_clip_train_classification_rad_3d(device):
         parameters["clip_mode"] = clip_mode
         parameters["nested_training"]["testing"] = -5
         parameters["nested_training"]["validation"] = -5
-        # shutil.rmtree(outputDir)  # overwrite previous results
-        Path(outputDir).mkdir(parents=True, exist_ok=True)
+        sanitize_outputDir()
         TrainingManager(
             dataframe=training_data,
             outputDir=outputDir,
@@ -657,7 +652,6 @@ def test_clip_train_classification_rad_3d(device):
             resume=False,
             reset=True,
         )
-    shutil.rmtree(outputDir)  # overwrite previous results
     print("passed")
 
 
@@ -688,6 +682,8 @@ def test_normtype_train_segmentation_rad_3d(device):
             parameters["model"]["norm_type"] = norm
             parameters["nested_training"]["testing"] = -5
             parameters["nested_training"]["validation"] = -5
+            if os.path.isdir(outputDir):
+                shutil.rmtree(outputDir)  # overwrite previous results
             Path(outputDir).mkdir(parents=True, exist_ok=True)
             TrainingManager(
                 dataframe=training_data,
@@ -697,7 +693,7 @@ def test_normtype_train_segmentation_rad_3d(device):
                 resume=False,
                 reset=True,
             )
-            shutil.rmtree(outputDir)  # overwrite previous results
+
         print("passed")
 
 
@@ -719,7 +715,7 @@ def test_metrics_segmentation_rad_2d(device):
     parameters["metrics"] = ["dice", "hausdorff", "hausdorff95"]
     parameters["model"]["architecture"] = "resunet"
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
@@ -728,7 +724,6 @@ def test_metrics_segmentation_rad_2d(device):
         resume=False,
         reset=True,
     )
-    shutil.rmtree(outputDir)  # overwrite previous results
 
     print("passed")
 
@@ -750,7 +745,7 @@ def test_metrics_regression_rad_2d(device):
     parameters["model"]["num_channels"] = 3
     parameters["model"]["architecture"] = "vgg11"
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
@@ -759,7 +754,6 @@ def test_metrics_regression_rad_2d(device):
         resume=False,
         reset=True,
     )
-    shutil.rmtree(outputDir)  # overwrite previous results
 
     print("passed")
 
@@ -787,7 +781,7 @@ def test_losses_segmentation_rad_2d(device):
         parameters["loss_function"] = loss_type
         parameters["nested_training"]["testing"] = -5
         parameters["nested_training"]["validation"] = -5
-        Path(outputDir).mkdir(parents=True, exist_ok=True)
+        sanitize_outputDir()
         TrainingManager(
             dataframe=training_data,
             outputDir=outputDir,
@@ -796,14 +790,14 @@ def test_losses_segmentation_rad_2d(device):
             resume=False,
             reset=True,
         )
-        shutil.rmtree(outputDir)  # overwrite previous results
+
     print("passed")
 
 
 def test_config_read():
     print("Starting testing reading configuration")
     # read and parse csv
-    file_config_temp = os.path.join(testingDir, "config_segmentation_temp.yaml")
+    file_config_temp = os.path.join(outputDir, "config_segmentation_temp.yaml")
     # if found in previous run, discard.
     if os.path.exists(file_config_temp):
         os.remove(file_config_temp)
@@ -897,7 +891,7 @@ def test_config_read():
 def test_cli_function_preprocess():
     print("Starting testing cli function preprocess")
     file_config = os.path.join(testingDir, "config_segmentation.yaml")
-    file_config_temp = os.path.join(testingDir, "config_segmentation_temp.yaml")
+    file_config_temp = os.path.join(outputDir, "config_segmentation_temp.yaml")
     # if found in previous run, discard.
     if os.path.exists(file_config_temp):
         os.remove(file_config_temp)
@@ -932,7 +926,8 @@ def test_cli_function_preprocess():
     # check that the length of training data is what we expect
     assert len(training_data) == 10, "Number of rows in dataframe is not 10"
 
-    shutil.rmtree(outputDir)  # overwrite previous results
+    if os.path.isdir(outputDir):
+        shutil.rmtree(outputDir)  # overwrite previous results
     print("passed")
 
 
@@ -941,7 +936,7 @@ def test_cli_function_mainrun(device):
     parameters = parseConfig(
         testingDir + "/config_segmentation.yaml", version_check_flag=False
     )
-    file_config_temp = os.path.join(testingDir, "config_segmentation_temp.yaml")
+    file_config_temp = os.path.join(outputDir, "config_segmentation_temp.yaml")
     # if found in previous run, discard.
     if os.path.exists(file_config_temp):
         os.remove(file_config_temp)
@@ -963,8 +958,11 @@ def test_cli_function_mainrun(device):
 
     file_data = os.path.join(inputDir, "train_2d_rad_segmentation.csv")
 
-    main_run(file_data, file_config_temp, outputDir, True, device, True)
-    shutil.rmtree(outputDir)  # overwrite previous results
+    main_run(
+        file_data, file_config_temp, outputDir, True, device, resume=False, reset=True
+    )
+    if os.path.isdir(outputDir):
+        shutil.rmtree(outputDir)  # overwrite previous results
     print("passed")
 
 
@@ -1005,7 +1003,7 @@ def test_dataloader_construction_train_segmentation_3d(device):
     parameters["weighted_loss"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
@@ -1014,7 +1012,7 @@ def test_dataloader_construction_train_segmentation_3d(device):
         resume=False,
         reset=True,
     )
-    shutil.rmtree(outputDir)  # overwrite previous results
+
     print("passed")
 
 
@@ -1157,7 +1155,7 @@ def test_checkpointing_segmentation_rad_2d(device):
     ]
     parameters["model"]["architecture"] = "unet"
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
@@ -1177,7 +1175,6 @@ def test_checkpointing_segmentation_rad_2d(device):
         resume=False,
         reset=False,
     )
-    shutil.rmtree(outputDir)  # overwrite previous results
 
     print("passed")
 
@@ -1297,7 +1294,11 @@ def test_anonymizer():
 
 def test_train_inference_segmentation_histology_2d(device):
     print("Starting histology train/inference tests")
-    output_dir_patches = os.path.join(testingDir, "histo_patches")
+    # overwrite previous results
+    if os.path.isdir(outputDir):
+        shutil.rmtree(outputDir)
+    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    output_dir_patches = os.path.join(outputDir, "histo_patches")
     if os.path.isdir(output_dir_patches):
         shutil.rmtree(output_dir_patches)
     Path(output_dir_patches).mkdir(parents=True, exist_ok=True)
@@ -1341,9 +1342,6 @@ def test_train_inference_segmentation_histology_2d(device):
     parameters["nested_training"]["testing"] = 1
     parameters["nested_training"]["validation"] = -2
     parameters["metrics"] = ["dice"]
-    # overwrite previous results
-    if os.path.isdir(outputDir):
-        shutil.rmtree(outputDir)
     Path(outputDir).mkdir(parents=True, exist_ok=True)
     TrainingManager(
         dataframe=training_data,
@@ -1357,14 +1355,11 @@ def test_train_inference_segmentation_histology_2d(device):
     inference_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_2d_histo_segmentation.csv", train=False
     )
-
     InferenceManager(
         dataframe=inference_data,
         outputDir=outputDir,
         parameters=parameters,
         device=device,
     )
-
-    shutil.rmtree(output_dir_patches)
 
     print("passed")
